@@ -1,8 +1,8 @@
-part of platform_maps_flutter;
+part of '../platform_maps_flutter.dart';
 
-typedef void MapCreatedCallback(PlatformMapController controller);
+typedef MapCreatedCallback = void Function(PlatformMapController controller);
 
-typedef void CameraPositionCallback(CameraPosition position);
+typedef CameraPositionCallback = void Function(CameraPosition position);
 
 class PlatformMap extends StatefulWidget {
   const PlatformMap({
@@ -164,16 +164,15 @@ class PlatformMap extends StatefulWidget {
   /// were not claimed by any other gesture recognizer.
   final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
   @override
-  _PlatformMapState createState() => _PlatformMapState();
+  State<PlatformMap> createState() => _PlatformMapState();
 }
 
 class _PlatformMapState extends State<PlatformMap> {
   @override
   Widget build(BuildContext context) {
     if (Platform.isAndroid) {
-      return googleMaps.GoogleMap(
-        initialCameraPosition:
-            widget.initialCameraPosition.googleMapsCameraPosition,
+      return google_maps.GoogleMap(
+        initialCameraPosition: widget.initialCameraPosition.googleMapsCameraPosition,
         compassEnabled: widget.compassEnabled,
         mapType: _getGoogleMapType(),
         padding: widget.padding,
@@ -196,13 +195,11 @@ class _PlatformMapState extends State<PlatformMap> {
         onTap: _onTap,
         onLongPress: _onLongPress,
         trafficEnabled: widget.trafficEnabled,
-        minMaxZoomPreference:
-            widget.minMaxZoomPreference.googleMapsZoomPreference,
+        minMaxZoomPreference: widget.minMaxZoomPreference.googleMapsZoomPreference,
       );
     } else if (Platform.isIOS) {
-      return appleMaps.AppleMap(
-        initialCameraPosition:
-            widget.initialCameraPosition.appleMapsCameraPosition,
+      return apple_maps.AppleMap(
+        initialCameraPosition: widget.initialCameraPosition.appleMapsCameraPosition,
         compassEnabled: widget.compassEnabled,
         mapType: _getAppleMapType(),
         padding: widget.padding,
@@ -224,11 +221,10 @@ class _PlatformMapState extends State<PlatformMap> {
         onTap: _onTap,
         onLongPress: _onLongPress,
         trafficEnabled: widget.trafficEnabled,
-        minMaxZoomPreference:
-            widget.minMaxZoomPreference.appleMapsZoomPreference,
+        minMaxZoomPreference: widget.minMaxZoomPreference.appleMapsZoomPreference,
       );
     } else {
-      return Text("Platform not yet implemented");
+      return const Text("Platform not yet implemented");
     }
   }
 
@@ -240,13 +236,13 @@ class _PlatformMapState extends State<PlatformMap> {
     if (Platform.isIOS) {
       widget.onCameraMove?.call(
         CameraPosition.fromAppleMapCameraPosition(
-          cameraPosition as appleMaps.CameraPosition,
+          cameraPosition as apple_maps.CameraPosition,
         ),
       );
     } else if (Platform.isAndroid) {
       widget.onCameraMove?.call(
         CameraPosition.fromGoogleMapCameraPosition(
-          cameraPosition as googleMaps.CameraPosition,
+          cameraPosition as google_maps.CameraPosition,
         ),
       );
     }
@@ -254,42 +250,39 @@ class _PlatformMapState extends State<PlatformMap> {
 
   void _onTap(dynamic position) {
     if (Platform.isIOS) {
-      widget.onTap?.call(LatLng._fromAppleLatLng(position as appleMaps.LatLng));
+      widget.onTap?.call(LatLng._fromAppleLatLng(position as apple_maps.LatLng));
     } else if (Platform.isAndroid) {
-      widget.onTap
-          ?.call(LatLng._fromGoogleLatLng(position as googleMaps.LatLng));
+      widget.onTap?.call(LatLng._fromGoogleLatLng(position as google_maps.LatLng));
     }
   }
 
   void _onLongPress(dynamic position) {
     if (Platform.isIOS) {
-      widget.onLongPress
-          ?.call(LatLng._fromAppleLatLng(position as appleMaps.LatLng));
+      widget.onLongPress?.call(LatLng._fromAppleLatLng(position as apple_maps.LatLng));
     } else if (Platform.isAndroid) {
-      widget.onLongPress
-          ?.call(LatLng._fromGoogleLatLng(position as googleMaps.LatLng));
+      widget.onLongPress?.call(LatLng._fromGoogleLatLng(position as google_maps.LatLng));
     }
   }
 
-  appleMaps.MapType _getAppleMapType() {
+  apple_maps.MapType _getAppleMapType() {
     if (widget.mapType == MapType.normal) {
-      return appleMaps.MapType.standard;
+      return apple_maps.MapType.standard;
     } else if (widget.mapType == MapType.satellite) {
-      return appleMaps.MapType.satellite;
+      return apple_maps.MapType.satellite;
     } else if (widget.mapType == MapType.hybrid) {
-      return appleMaps.MapType.hybrid;
+      return apple_maps.MapType.hybrid;
     }
-    return appleMaps.MapType.standard;
+    return apple_maps.MapType.standard;
   }
 
-  googleMaps.MapType _getGoogleMapType() {
+  google_maps.MapType _getGoogleMapType() {
     if (widget.mapType == MapType.normal) {
-      return googleMaps.MapType.normal;
+      return google_maps.MapType.normal;
     } else if (widget.mapType == MapType.satellite) {
-      return googleMaps.MapType.satellite;
+      return google_maps.MapType.satellite;
     } else if (widget.mapType == MapType.hybrid) {
-      return googleMaps.MapType.hybrid;
+      return google_maps.MapType.hybrid;
     }
-    return googleMaps.MapType.normal;
+    return google_maps.MapType.normal;
   }
 }
